test: consolidate equalizer preamp coverage into one focused test
